Enhancement: Rework error/information logging
Error and information logging is reworked so that the target database name, target instance number and target database id is always logged in a standardized way.

Previously only the target database name was logged within the message text itself which made it hard to track down error and information messages in RAC environments or if multiple target databases had the same database name.